Move metadata manipulation to one spot.
