+++ CMakeLists.txt
@@ -60,12 +60,13 @@ set(codegen_scripts
   "${codegen_dir}/common_helpers.py"
   "${codegen_dir}/generate_service.py"
   "${codegen_dir}/service_helpers.py"
+  "${codegen_dir}/proto_helpers.py"
+  "${codegen_dir}/metadata_mutation.py"
   "${codegen_dir}/templates/library_interface.h.mako"
   "${codegen_dir}/templates/library.cpp.mako"
   "${codegen_dir}/templates/library.h.mako"
   "${codegen_dir}/templates/mock_library.h.mako"
   "${codegen_dir}/templates/proto.mako"
-  "${codegen_dir}/proto_helpers.py"
   "${codegen_dir}/templates/service.cpp.mako"
   "${codegen_dir}/templates/service_helpers.mako"
   "${codegen_dir}/templates/service.h.mako")
