+++ source/codegen/generate_service.py
@@ -5,6 +5,7 @@ import importlib
 import importlib.util
 import mako.template
 import pathlib
+import metadata_mutation
 from mako.lookup import TemplateLookup
 
 def generate_service_file(metadata, template_file_name, generated_file_suffix, gen_dir):
@@ -23,6 +24,12 @@ def generate_service_file(metadata, template_file_name, generated_file_suffix, g
   f.write(template.render(data=metadata))
   f.close()
 
+def fix_up_metadata(metadata):
+  for function in metadata["functions"]:
+    parameters = metadata["functions"][function]["parameters"]
+    metadata_mutation.sanitize_names(parameters)
+    metadata_mutation.mark_non_grpc_params(parameters)
+
 def generate_all(metadata_dir, gen_dir):
   sys.path.append(metadata_dir)
   init_file = os.path.join(metadata_dir, "__init__.py")
@@ -33,6 +40,7 @@ def generate_all(metadata_dir, gen_dir):
   metadata = module.metadata
   lookup = TemplateLookup(directories = metadata_dir)
   metadata["lookup"] = lookup
+  fix_up_metadata(metadata)
   generate_service_file(metadata, "proto.mako", ".proto", gen_dir)
   generate_service_file(metadata, "service.h.mako", "_service.h", gen_dir)
   generate_service_file(metadata, "service.cpp.mako", "_service.cpp", gen_dir)
