+++ source/codegen/metadata_mutation.py
@@ -1,0 +1,41 @@
+import common_helpers
+
+RESERVED_WORDS = [
+    'abstract', 'as',
+    'base', 'bool', 'break','byte',
+    'case', 'catch', 'char', 'checked', 'class', 'const', 'continue',
+    'decimal', 'default', 'delegate', 'do', 'double',
+    'else', 'enum', 'event', 'explicit', 'extern',
+    'false', 'finally', 'fixed', 'float', 'for', 'foreach',
+    'goto',
+    'if', 'implicit', 'in', 'int', 'interface', 'internal', 'is',
+    'lock', 'long',
+    'namespace', 'new', 'null',
+    'object', 'operator', 'out', 'override',
+    'params', 'private', 'protected', 'public',
+    'readonly', 'ref', 'return',
+    'sbyte', 'sealed', 'short', 'sizeof', 'stackalloc', 'static', 'status', 'string', 'struct', 'switch',
+    'this', 'throw', 'true', 'try', 'typeof',
+    'uint', 'ulong', 'unchecked', 'unsafe', 'ushort', 'using',
+    'virtual', 'void', 'volatile',
+    'while'
+]
+
+def sanitize_names(parameters):
+    """Sanitizes name fields on a list of parameter objects and populates the cppname field with the sanitized value."""
+    for parameter in parameters:
+        parameter['cppName'] = parameter['name']
+        if parameter['cppName'] in RESERVED_WORDS:
+            parameter['cppName'] = parameter['cppName'] + 'Parameter'
+
+def mark_non_grpc_params(parameters):
+  named_params = { p['name'] : p for p in parameters }
+  for param in parameters:
+    mechanism = common_helpers.get_size_mechanism(param)
+    if mechanism in {'len', 'ivi-dance', 'passed-in'}:
+      size_param = named_params.get(param['size']['value'], None)
+      size_param['is_size_param'] = True
+      if mechanism == 'len' or mechanism == 'ivi-dance':
+        size_param['include_in_proto'] = False
+        if mechanism == 'len':
+          size_param['determine_size_from'] = param['name']
